Rearrange constants files
