+++ apps/web/app/(ee)/api/partner-profile/programs/[programId]/events/route.ts
@@ -2,8 +2,8 @@ import { getEvents } from "@/lib/analytics/get-events";
 import { DubApiError } from "@/lib/api/errors";
 import { getProgramEnrollmentOrThrow } from "@/lib/api/programs/get-program-enrollment-or-throw";
 import { withPartnerProfile } from "@/lib/auth/partner";
+import { LARGE_PROGRAM_MIN_TOTAL_COMMISSIONS_CENTS } from "@/lib/constants/program";
 import { generateRandomName } from "@/lib/names";
-import { LARGE_PROGRAM_MIN_TOTAL_COMMISSIONS_CENTS } from "@/lib/partners/constants";
 import {
   PartnerProfileLinkSchema,
   partnerProfileEventsQuerySchema,
