+++ apps/web/app/(ee)/api/stripe/webhook/utils/process-payout-invoice-failure.ts
@@ -1,7 +1,7 @@
 import {
   DIRECT_DEBIT_PAYMENT_METHOD_TYPES,
   PAYOUT_FAILURE_FEE_CENTS,
-} from "@/lib/partners/constants";
+} from "@/lib/constants/payouts";
 import { createPaymentIntent } from "@/lib/stripe/create-payment-intent";
 import { sendBatchEmail } from "@dub/email";
 import PartnerPayoutFailed from "@dub/email/templates/partner-payout-failed";
