+++ apps/web/app/(ee)/partners.dub.co/(dashboard)/programs/[programSlug]/(enrolled)/customers/[customerId]/page-client.tsx
@@ -1,6 +1,6 @@
 "use client";
 
-import { CUSTOMER_PAGE_EVENTS_LIMIT } from "@/lib/partners/constants";
+import { CUSTOMER_PAGE_EVENTS_LIMIT } from "@/lib/constants/misc";
 import usePartnerCustomer from "@/lib/swr/use-partner-customer";
 import useProgramEnrollment from "@/lib/swr/use-program-enrollment";
 import { PartnerEarningsResponse } from "@/lib/types";
