+++ apps/web/app/app.dub.co/(dashboard)/[slug]/(ee)/program/partners/import-export-buttons.tsx
@@ -1,6 +1,6 @@
 "use client";
 
-import { PROGRAM_IMPORT_SOURCES } from "@/lib/partners/constants";
+import { PROGRAM_IMPORT_SOURCES } from "@/lib/constants/program";
 import useWorkspace from "@/lib/swr/use-workspace";
 import { useExportPartnersModal } from "@/ui/modals/export-partners-modal";
 import { useImportFirstPromoterModal } from "@/ui/modals/import-firstpromoter-modal";
