+++ apps/web/lib/actions/partners/create-bounty-submission.ts
@@ -4,10 +4,12 @@ import { createId } from "@/lib/api/create-id";
 import { getWorkspaceUsers } from "@/lib/api/get-workspace-users";
 import { getProgramEnrollmentOrThrow } from "@/lib/api/programs/get-program-enrollment-or-throw";
 import {
+  BOUNTY_MAX_SUBMISSION_DESCRIPTION_LENGTH,
+  BOUNTY_MAX_SUBMISSION_FILES,
+  BOUNTY_MAX_SUBMISSION_URLS,
+} from "@/lib/constants/bounties";
+import {
   BountySubmissionFileSchema,
-  MAX_BOUNTY_SUBMISSION_DESCRIPTION_LENGTH,
-  MAX_SUBMISSION_FILES,
-  MAX_SUBMISSION_URLS,
   submissionRequirementsSchema,
 } from "@/lib/zod/schemas/bounties";
 import { sendBatchEmail, sendEmail } from "@dub/email";
@@ -25,13 +27,13 @@ const schema = z.object({
   bountyId: z.string(),
   files: z
     .array(BountySubmissionFileSchema)
-    .max(MAX_SUBMISSION_FILES)
+    .max(BOUNTY_MAX_SUBMISSION_FILES)
     .default([]),
-  urls: z.array(z.string().url()).max(MAX_SUBMISSION_URLS).default([]),
+  urls: z.array(z.string().url()).max(BOUNTY_MAX_SUBMISSION_URLS).default([]),
   description: z
     .string()
     .trim()
-    .max(MAX_BOUNTY_SUBMISSION_DESCRIPTION_LENGTH)
+    .max(BOUNTY_MAX_SUBMISSION_DESCRIPTION_LENGTH)
     .optional(),
   isDraft: z
     .boolean()
