+++ apps/web/lib/constants/partner-profile.ts
@@ -1,0 +1,2 @@
+export const MAX_INVITES_PER_REQUEST = 5;
+export const MAX_PARTNER_USERS = 10;
